setting up scenarios
